# workshop-4/python-temp/swagger.yml
openapi: 3.1.0
info:
  title: LBK Points - Transfer API
  version: 1.1.0
  description: |
    โอนแต้ม, ดูสถานะ, และค้นประวัติ พร้อม CRUD User management
servers:
  - url: http://localhost:3000
...
components:
  schemas:
...
    TransferStatus:
      type: string
      enum: [pending, processing, completed, failed, cancelled, reversed]
...
    Transfer:
      type: object
      required:
        - idemKey
        - fromUserId
        - toUserId
        - amount
        - status
        - createdAt
        - updatedAt
      properties:
        idemKey:
          type: string
          description: Idempotency key as main reference ID
        transferId:
          type: integer
          nullable: true
          description: Internal system ID (autoincrement)
        fromUserId:
          type: integer
          minimum: 1
          description: Sender user ID
        toUserId:
          type: integer
          minimum: 1
          description: Receiver user ID
        amount:
          type: integer
          minimum: 1
          description: Transfer amount
        status:
          $ref: '#/components/schemas/TransferStatus'
        note:
          type: string
          nullable: true
          maxLength: 512
          description: Transfer note
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Completion timestamp
        failReason:
          type: string
          nullable: true
          description: Failure reason if failed
    TransferCreateRequest:
      type: object
      required:
        - fromUserId
        - toUserId
        - amount
      properties:
        fromUserId:
          type: integer
          minimum: 1
          description: Sender user ID
        toUserId:
          type: integer
          minimum: 1
          description: Receiver user ID
        amount:
          type: integer
          minimum: 1
          description: Amount to transfer
        note:
          type: string
          nullable: true
          maxLength: 512
          description: Transfer note
    TransferCreateResponse:
      type: object
      properties:
        transfer:
          $ref: '#/components/schemas/Transfer'
    TransferGetResponse:
      type: object
      properties:
        transfer:
          $ref: '#/components/schemas/Transfer'
    TransferListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transfer'
        page:
          type: integer
          minimum: 1
          description: Current page number
        pageSize:
          type: integer
          minimum: 1
          description: Items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: amount must be > 0
        details:
          type: object
          nullable: true
          description: Additional error details
paths:
...
  /transfers:
    post:
      tags: [Transfers]
      summary: สร้างคำสั่งโอนแต้ม (ระบบจะสร้าง Idempotency-Key ให้)
      description: |
        สร้างรายการโอนแต้มแบบอะตอมมิก ระบบจะ generate `idemKey` (Idempotency-Key)
        และคืนค่าไว้ใช้ติดตามสถานะผ่าน GET /transfers/{id}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCreateRequest'
            example:
              fromUserId: 101
              toUserId: 202
              amount: 250
              note: ขอบคุณสำหรับความช่วยเหลือ
      responses:
        '201':
          description: สร้างสำเร็จ
          headers:
            Idempotency-Key:
              description: idemKey ที่ระบบสร้างให้ (ใช้เป็น /transfers/{id})
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferCreateResponse'
              example:
                transfer:
                  idemKey: "5d1f8c7a-2b5b-4b1f-9f2a-8f50b0a8d9f3"
                  transferId: 9876
                  fromUserId: 101
                  toUserId: 202
                  amount: 250
                  status: completed
                  note: "ขอบคุณสำหรับความช่วยเหลือ"
                  createdAt: "2025-10-16T14:03:12Z"
                  updatedAt: "2025-10-16T14:03:12Z"
                  completedAt: "2025-10-16T14:03:12Z"
        '400':
          description: คำขอไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: ความขัดแย้ง (เช่น แต้มไม่พอ)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: ตรวจรูปแบบผ่าน แต่ทำงานต่อไม่ได้ (เช่น โอนให้ตัวเอง)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    get:
      tags: [Transfers]
      summary: ค้น/ดูประวัติการโอน (กรองด้วย userId เท่านั้น)
      description: แสดงรายการที่ userId เกี่ยวข้อง (ทั้ง sender และ receiver) พร้อมแบ่งหน้า
      parameters:
        - name: userId
          in: query
          required: true
          description: แสดงเฉพาะรายการที่เกี่ยวข้องกับ userId (ทั้งโอนออกและรับเข้า)
          schema:
            type: integer
            minimum: 1
        - name: page
          in: query
          required: false
          description: หน้าที่ต้องการ (เริ่มที่ 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          description: จำนวนต่อหน้า (1–200)
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: รายการที่พบ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferListResponse'
              example:
                data:
                  - idemKey: "5d1f8c7a-2b5b-4b1f-9f2a-8f50b0a8d9f3"
                    transferId: 9876
                    fromUserId: 101
                    toUserId: 202
                    amount: 250
                    status: completed
                    createdAt: "2025-10-16T14:03:12Z"
                    updatedAt: "2025-10-16T14:03:12Z"
                    completedAt: "2025-10-16T14:03:12Z"
                  - idemKey: "a8b4f2e0-5562-4f1c-9b62-2a2f2f4c9b10"
                    transferId: 9901
                    fromUserId: 303
                    toUserId: 101
                    amount: 100
                    status: completed
                    createdAt: "2025-10-15T10:00:00Z"
                    updatedAt: "2025-10-15T10:00:00Z"
                    completedAt: "2025-10-15T10:00:00Z"
                page: 1
                pageSize: 20
                total: 2
        '400':
          description: คำขอไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /transfers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Transfer ID สำหรับค้นหาสถานะ (เท่ากับ Idempotency-Key ที่ระบบสร้างให้ตอน POST /transfers)
        schema:
          type: string
          minLength: 8
          maxLength: 128
    
    get:
      tags: [Transfers]
      summary: ดูสถานะคำสั่งโอน (ใช้ idemKey เป็น id)
      responses:
        '200':
          description: ข้อมูลรายการโอน
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferGetResponse'
              example:
                transfer:
                  idemKey: "5d1f8c7a-2b5b-4b1f-9f2a-8f50b0a8d9f3"
                  transferId: 9876
                  fromUserId: 101
                  toUserId: 202
                  amount: 250
                  status: completed
                  note: "ขอบคุณสำหรับช่วยงาน"
                  createdAt: "2025-10-16T14:03:12Z"
                  updatedAt: "2025-10-16T14:03:12Z"
                  completedAt: "2025-10-16T14:03:12Z"
        '404':
          description: ไม่พบข้อมูล
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'